Multiple actions
Installed font-awsome and added icons for delete and check the tasks. Implemented the fixFooter function to change position of the footer if the pages include vertical scroll. Installed moment Js and appllied to the tasks generating the messages about how long ago they were created. Fixed the navbar to always be visible to include more tasks even if the list is long and the user scrolled to the bottom of the list. Created the app icons in app/platform/android/icons for different android devices sizes.